+++ setup.py
@@ -12,8 +12,8 @@ setup(
     license='MIT',
     author='SvenMatzke',
     author_email='matzke.sven@googlemail.com',
-    description='Minimal socket webserver for designed to work on esp8266 and up with realy small footprint',
+    description='Minimal webserver for micropython Designed to work on esp8266 for socketserving and async for esp32',
     long_description=README,
     cmdclass={'sdist': sdist_upip.sdist},
-    install_requires=['ujson']
+    install_requires=['ujson', 'micropython-ulogging']
 )
